+++ packages/analytics-js-plugins/src/ketchConsentManager/index.ts
@@ -7,11 +7,11 @@ import type { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import type { DestinationConfig } from '@rudderstack/analytics-js-common/types/Destination';
 import type { IErrorHandler } from '@rudderstack/analytics-js-common/types/ErrorHandler';
 import type { PluginName } from '@rudderstack/analytics-js-common/types/PluginsManager';
-import { checks } from '../shared-chunks/common';
 import { DESTINATION_CONSENT_STATUS_ERROR } from './logMessages';
 import { KETCH_CONSENT_MANAGER_PLUGIN } from './constants';
 import type { KetchConsentData } from './types';
 import { getKetchConsentData, updateConsentStateFromData } from './utils';
+import { isUndefined } from '../shared-chunks/common';
 
 const pluginName: PluginName = 'KetchConsentManager';
 
@@ -47,7 +47,7 @@ const KetchConsentManager = (): ExtensionPlugin => ({
     ): void {
       // retrieve consent data and update the state
       let ketchConsentData;
-      if (!checks.isUndefined((globalThis as any).ketchConsent)) {
+      if (!isUndefined((globalThis as any).ketchConsent)) {
         ketchConsentData = (globalThis as any).ketchConsent;
       } else {
         ketchConsentData = getKetchConsentData(storeManager, logger);