+++ packages/analytics-js-integrations/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.11.8](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.11.7...@rudderstack/analytics-js-integrations@3.11.8) (2024-11-22)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.9`
 ## [3.11.7](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.11.6...@rudderstack/analytics-js-integrations@3.11.7) (2024-11-22)
 
 ### Dependency Updates