+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.6.13](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.12...@rudderstack/analytics-js-plugins@3.6.13) (2024-11-22)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.9`
+* `@rudderstack/analytics-js` updated to version `3.11.9`
+* `@rudderstack/analytics-js-cookies` updated to version `0.4.11`
+
+### Bug Fixes
+
+* reorganize common imports in plugins ([bc569ba](https://github.com/rudderlabs/rudder-sdk-js/commit/bc569baa4ae2211696934d489a370e2e9ca80521))
+* restore data sanitization changes ([2a13e7c](https://github.com/rudderlabs/rudder-sdk-js/commit/2a13e7c463b2d480f6d9a23f32abe4e56f6557d4))
+
 ## [3.6.12](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.11...@rudderstack/analytics-js-plugins@3.6.12) (2024-11-22)
 
 ### Dependency Updates